+++ lib/routes/interactionRoute.js
@@ -23,6 +23,9 @@ module.exports = router
 
   //This route should handle all accept/reject/report votes.
   .post('/', bodyParser, (req, res, next)=>{
+    // console.log(req.body);
+    //TODO: decide what report should do and implement it as an if block here.
+
     new Interaction(req.body).save()
       .then(saved => {
         // vote = saved;
