+++ public/index.html
@@ -143,9 +143,7 @@
       </div>
     </div>
     <script src="https://code.jquery.com/jquery-3.1.0.min.js" integrity="sha256-cCueBR6CsyA4/9szpPfrX3s49M9vUU5BgtiJj06wt/s=" crossorigin="anonymous"></script>
-    <script type="text/javascript" src="/js/view.js"></script>
+    <script type="text/javascript" src="/js/index.js"></script>
     <script type="text/javascript" src="/js/login.js"></script>
-    <script type="text/javascript" src="/js/signup.js"></script>
-    <script type="text/javascript" src="/js/vote.js"></script>
   </body>
 </html>
